promoGroup-scraping/requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.3.0
h11==0.14.0
idna==3.4
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.6
